refactor(file): move storage and naming services to services package; update imports and usages
