+++ file/src/main/java/pt/estga/file/services/application/MediaService.java
@@ -1,4 +1,4 @@
-package pt.estga.file.application;
+package pt.estga.file.services.application;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
