+++ file/src/main/java/pt/estga/file/services/naming/FileNamingService.java
@@ -1,4 +1,4 @@
-package pt.estga.file.naming;
+package pt.estga.file.services.naming;
 
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
