+++ file/src/main/java/pt/estga/file/services/naming/StoragePathStrategy.java
@@ -1,4 +1,4 @@
-package pt.estga.file.naming;
+package pt.estga.file.services.naming;
 
 import org.springframework.stereotype.Component;
 import pt.estga.file.entities.MediaFile;
