+++ file/src/main/java/pt/estga/file/services/storage/FileStorageServiceLocalImpl.java
@@ -1,4 +1,4 @@
-package pt.estga.file.storage;
+package pt.estga.file.services.storage;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
